% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_sums.R
\name{get_col_sums_added}
\alias{get_col_sums_added}
\title{get column sums added to a dataframe}
\usage{
get_col_sums_added(a_df, desc_of_row = "Total")
}
\arguments{
\item{a_df}{a dataframe;}

\item{desc_of_row}{description of the row,
and the default value is 'Total'}
}
\description{
This function returns a new dataframe with its last row being the
column sums of the original dataframe. Note that the first column of the
input dataframe must be of character type.
}
\examples{
df <-
  data.frame(x = letters[1:10],
             y = 1,
             z = 2,
             w = rnorm(10))

df_with_col_sums <- get_col_sums_added(df)
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{column}
\keyword{dataframe;}
\keyword{sums}
