% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_concordance.R
\name{make_concordance}
\alias{make_concordance}
\title{make a concordance using two variables of a dataframe}
\usage{
make_concordance(a_df, v1, v2)
}
\arguments{
\item{a_df:}{a dataframe;}

\item{v1:}{a varaible of a_df;}

\item{v2:}{another variable of a_df}
}
\description{
This function makes a dataframe, which is for a concordance
}
\examples{
a_df <-
  data.frame(x = sample(letters)) \%>\%
  mutate(y = toupper(x),
         z = rnorm(26))
concordance_df <- make_concordance(a_df, "x", "y")
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{concordance;}
\keyword{dataframe}
