% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_max.R
\name{modified_max}
\alias{modified_max}
\title{modified max function}
\usage{
modified_max(an_integer_vec)
}
\arguments{
\item{an_integer_vec}{a vector}
}
\description{
if an_integer_vec has only NA's then return 0, otherwise return max(x, na.rm = TRUE)
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{max}
\keyword{modidied}
