% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_min.R
\name{modified_min}
\alias{modified_min}
\title{modified min function}
\usage{
modified_min(an_integer_vec)
}
\arguments{
\item{an_integer_vec}{a vector}
}
\description{
if an_integer_vec has only NA's then return 0, otherwise return min(x, na.rm = TRUE)
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{min}
\keyword{modidied}
