% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_df_key_info.R
\name{find_df_key_info}
\alias{find_df_key_info}
\title{find key information about a dataframe}
\usage{
find_df_key_info(a_df)
}
\arguments{
\item{a_df}{a dataframe}
}
\description{
This function allows you to find key information about a dataframe
}
\examples{
find_df_key_info()
}
\keyword{dataframe}
\keyword{info;}
\keyword{key}
